refactor: address PR review — explicit no-op catch, shared localStorage test mock
